+++ CHANGELOG.md
@@ -21,6 +21,11 @@ For each release, changes are grouped under these headings:
 
 The following changes have been completed, and will be included in the next tagged release.
 
+### New
+
+* Content Types
+  - added `ContentType` branded string
+
 ## v1.2.1
 
 Released Thursday, 28th May 2020.